+++ app/src/main/java/jp/ac/metro_cit/adv_prog_2024/gomoku/interfaces/SocketProps.java
@@ -1,4 +1,4 @@
-package jp.ac.metro_cit.adv_prog_2024.gomoku.communication;
+package jp.ac.metro_cit.adv_prog_2024.gomoku.interfaces;
 
 /**
  * ソケット起動時の引数を格納するためのインターフェイス
