+++ app/src/main/java/jp/ac/metro_cit/adv_prog_2024/gomoku/models/GameState.java
@@ -1,3 +1,5 @@
 package jp.ac.metro_cit.adv_prog_2024.gomoku.models;
 
-public class GameState {}
+import java.io.Serializable;
+
+public record GameState(String data) implements Serializable {}
